Fix bind parameter address instability issue
The bind parameters rely on stable addresses in memory, however the Vec
used to hold them was being allowed to re-allocate and therefore change
the addresses refered to by the OCI execute function. Now the Vec is
reserved with the right number of elements to ensure that it does not get
increases automatically.